refactoring boundary updates in FieldBlock; fixing / commenting memory leaks; split face_[lower|upper]_[xyz] into separate face_* and axis_*, adding old face_enum as face_axis_enum (for Tree code); fixed bug introduced in r2009-r2010 where deleted pointer was not set to 0; cleaned enzo-p.cpp; refactoring EnzoSimulation; updated SConstruct for charm + papi

+++ src/Mesh/mesh_ItBlocks.hpp
@@ -35,7 +35,7 @@ private: // attributes
   Patch * patch_;
 
   /// Index to the current DataBlock
-  int curr_;
+  size_t curr_;
 };
 
 #endif /* MESH_IT_BLOCKS_HPP */
